// custom-code-region-ts-oop-example/src/models/operations/collectionadd.ts
export type CollectionAddRequest = {
  tenant?: string | undefined;
  databaseName?: string | undefined;
  collectionId: string;
  addCollectionRecordsPayload: components.AddCollectionRecordsPayload;
};