// custom-code-region-ts-oop-example/src/models/operations/deletecollection.ts
export type DeleteCollectionRequest = {
  /**
   * Tenant ID
   */
  tenantId?: string | undefined;
  /**
   * Database name
   */
  databaseName?: string | undefined;
  /**
   * UUID of the collection to delete
   */
  collectionId: string;
};