// custom-code-region-ts-oop-example/src/models/operations/updatecollection.ts
export type UpdateCollectionRequest = {
  /**
   * Tenant ID
   */
  tenantId?: string | undefined;
  /**
   * Database name
   */
  databaseName?: string | undefined;
  /**
   * UUID of the collection to update
   */
  collectionId: string;
  updateCollectionPayload: components.UpdateCollectionPayload;
};